+++ modules/datastudio/google-jwt-client.ts
@@ -1,4 +1,4 @@
-import { google } from 'googleapis';
+import * as google from 'googleapis-common';
 import { env } from '../../app/env';
 import { JWT } from 'google-auth-library';
 import { networkContext } from '../network/network-context.service';
@@ -6,13 +6,13 @@ import { DeploymentEnv } from '../network/network-config-types';
 
 export class GoogleJwtClient {
     public async getAuthorizedSheetsClient(privateKey: string): Promise<JWT> {
-        const jwtClient = new google.auth.JWT(
+        const jwtClient = new google.JWT(
             networkContext.data.datastudio![env.DEPLOYMENT_ENV as DeploymentEnv].user,
             undefined,
             privateKey,
             'https://www.googleapis.com/auth/spreadsheets',
         );
-        await jwtClient.authorize(function (err, result) {
+        jwtClient.authorize(function (err, result) {
             if (err) {
                 console.log(`Error authorizing google jwt client: ${err}`);
             }