+++ modules/pool/lib/pool-gql-loader.service.ts
@@ -34,15 +34,17 @@ import {
     GqlPoolAprItemType,
     GqlUserStakedBalance,
     GqlPoolFilterCategory,
+    HookData,
 } from '../../../schema';
 import { isSameAddress } from '@balancer-labs/sdk';
-import _, { has } from 'lodash';
+import _, { has, map } from 'lodash';
 import { prisma } from '../../../prisma/prisma-client';
 import {
     Chain,
     Prisma,
     PrismaPoolAprType,
     PrismaPriceRateProviderData,
+    PrismaTokenTypeOption,
     PrismaUserStakedBalance,
     PrismaUserWalletBalance,
 } from '@prisma/client';
@@ -54,6 +56,7 @@ import { GithubContentService } from '../../content/github-content.service';
 import { SanityContentService } from '../../content/sanity-content.service';
 import { ElementData, FxData, GyroData, StableData } from '../subgraph-mapper';
 import { ZERO_ADDRESS } from '@balancer/sdk';
+import { tokenService } from '../../token/token.service';
 
 export class PoolGqlLoaderService {
     public async getPool(id: string, chain: Chain, userAddress?: string): Promise<GqlPoolUnion> {
@@ -80,11 +83,57 @@ export class PoolGqlLoaderService {
         );
 
         // load rate provider data into PoolTokenDetail model
+        await this.enrichWithRateproviderData(mappedPool);
+
+        // load underlying token info into PoolTokenDetail and GqlPoolTokenDisplay
+        await this.enrichWithUnderlyingTokenData(mappedPool);
+
+        return mappedPool;
+    }
+
+    private async enrichWithUnderlyingTokenData(mappedPool: GqlPoolUnion) {
+        for (const token of mappedPool.poolTokens) {
+            if (token.isErc4626) {
+                const prismaToken = await prisma.prismaToken.findUnique({
+                    where: { address_chain: { address: token.address, chain: mappedPool.chain } },
+                });
+                if (prismaToken?.underlyingTokenAddress) {
+                    const underlyingTokenDefinition = await tokenService.getTokenDefinition(
+                        prismaToken.underlyingTokenAddress,
+                        mappedPool.chain,
+                    );
+                    token.underlyingToken = underlyingTokenDefinition;
+                    mappedPool.displayTokens.push();
+                }
+            }
+            if (token.hasNestedPool) {
+                for (const nestedToken of token.nestedPool!.tokens) {
+                    if (nestedToken.isErc4626) {
+                        const prismaToken = await prisma.prismaToken.findUnique({
+                            where: { address_chain: { address: nestedToken.address, chain: mappedPool.chain } },
+                        });
+                        if (prismaToken?.underlyingTokenAddress) {
+                            const tokenDefinition = await tokenService.getTokenDefinition(
+                                prismaToken.underlyingTokenAddress,
+                                mappedPool.chain,
+                            );
+                            nestedToken.underlyingToken = tokenDefinition;
+                        }
+                    }
+                }
+            }
+        }
+    }
+
+    private async enrichWithRateproviderData(mappedPool: GqlPoolUnion) {
         for (const token of mappedPool.poolTokens) {
             if (token.priceRateProvider && token.priceRateProvider !== ZERO_ADDRESS) {
                 const rateproviderData = await prisma.prismaPriceRateProviderData.findUnique({
                     where: {
-                        chain_rateProviderAddress: { chain: chain, rateProviderAddress: token.priceRateProvider },
+                        chain_rateProviderAddress: {
+                            chain: mappedPool.chain,
+                            rateProviderAddress: token.priceRateProvider,
+                        },
                     },
                 });
                 if (rateproviderData) {
@@ -107,7 +156,7 @@ export class PoolGqlLoaderService {
                         const rateproviderData = await prisma.prismaPriceRateProviderData.findUnique({
                             where: {
                                 chain_rateProviderAddress: {
-                                    chain: chain,
+                                    chain: mappedPool.chain,
                                     rateProviderAddress: nestedToken.priceRateProvider,
                                 },
                             },
@@ -128,8 +177,6 @@ export class PoolGqlLoaderService {
                 }
             }
         }
-
-        return mappedPool;
     }
 
     public async getPools(args: QueryPoolGetPoolsArgs): Promise<GqlPoolMinimal[]> {
@@ -192,6 +239,13 @@ export class PoolGqlLoaderService {
     ): GqlPoolMinimal {
         return {
             ...pool,
+            hook:
+                (pool.hook &&
+                    pool.hook.dynamicData && {
+                        ...pool.hook,
+                        dynamicData: pool.hook.dynamicData as HookData,
+                    }) ||
+                undefined,
             incentivized: pool.categories.some((category) => category === 'INCENTIVIZED'),
             vaultVersion: pool.protocolVersion,
             decimals: 18,
@@ -202,6 +256,7 @@ export class PoolGqlLoaderService {
             userBalance: this.getUserBalance(pool, userWalletbalances, userStakedBalances),
             categories: pool.categories as GqlPoolFilterCategory[],
             tags: pool.categories,
+            hasErc4626: pool.allTokens.some((token) => token.token.types.some((type) => type.type === 'ERC4626')),
         };
     }
 
@@ -510,6 +565,13 @@ export class PoolGqlLoaderService {
             vaultVersion: poolWithoutTypeData.protocolVersion,
             categories: pool.categories as GqlPoolFilterCategory[],
             tags: pool.categories,
+            hook:
+                (pool.hook &&
+                    pool.hook.dynamicData && {
+                        ...pool.hook,
+                        dynamicData: pool.hook.dynamicData as HookData,
+                    }) ||
+                undefined,
         };
 
         //TODO: may need to build out the types here still
@@ -582,7 +644,10 @@ export class PoolGqlLoaderService {
             const poolToken = pool.tokens.find((poolToken) => poolToken.address === token.token.address);
             const isNested = !poolToken;
             const isPhantomBpt = token.tokenAddress === pool.address;
-            const isMainToken = !token.token.types.some((type) => type.type === 'PHANTOM_BPT' || type.type === 'BPT');
+            const isMainToken = !token.token.types.some(
+                (type) => type.type === 'PHANTOM_BPT' || type.type === 'BPT' || type.type === 'ERC4626',
+            );
+            const isErc4626 = token.token.types.some((type) => type.type === 'ERC4626');
 
             return {
                 ...token.token,
@@ -591,6 +656,7 @@ export class PoolGqlLoaderService {
                 isNested,
                 isPhantomBpt,
                 isMainToken,
+                isErc4626,
             };
         });
     }
@@ -600,7 +666,7 @@ export class PoolGqlLoaderService {
             .filter((token) => token.address !== pool.address)
             .map((poolToken) => {
                 const allToken = pool.allTokens.find((allToken) => allToken.token.address === poolToken.address);
-                if (allToken?.nestedPool?.type === 'COMPOSABLE_STABLE') {
+                if (allToken?.nestedPool) {
                     const mainTokens =
                         allToken.nestedPool.allTokens.filter(
                             (nestedToken) =>
@@ -645,6 +711,7 @@ export class PoolGqlLoaderService {
             isAllowed: poolToken.token.types.some(
                 (type) => type.type === 'WHITE_LISTED' || type.type === 'PHANTOM_BPT' || type.type === 'BPT',
             ),
+            isErc4626: poolToken.token.types.some((type) => type.type === 'ERC4626'),
         };
     }
 
@@ -1054,12 +1121,15 @@ export class PoolGqlLoaderService {
                     apr: aprItem.range.min,
                     type: type,
                 });
-                aprItems.push({
-                    id: aprItem.id,
-                    title: aprItem.title,
-                    apr: aprItem.range.max - aprItem.range.min,
-                    type: 'STAKING_BOOST',
-                });
+                if (pool.type !== 'COW_AMM') {
+                    // TODO remove this when we have a better solution
+                    aprItems.push({
+                        id: `${aprItem.id}-boost`,
+                        title: aprItem.title,
+                        apr: aprItem.range.max - aprItem.range.min,
+                        type: 'STAKING_BOOST',
+                    });
+                }
             } else {
                 aprItems.push({
                     id: aprItem.id,
@@ -1245,6 +1315,7 @@ export class PoolGqlLoaderService {
                     },
                 },
                 userWalletBalances: false,
+                hook: true,
             };
         }
 
@@ -1273,6 +1344,7 @@ export class PoolGqlLoaderService {
                     balanceNum: { gt: 0 },
                 },
             },
+            hook: true,
         };
     }
 }