+++ modules/sor/sor-debug.test.ts
@@ -1,15 +1,17 @@
+// yarn vitest sor-debug.test.ts
+import { Chain } from '@prisma/client';
 import { initRequestScopedContext, setRequestScopedContextValue } from '../context/request-scoped-context';
-import { poolService } from '../pool/pool.service';
+import { chainIdToChain } from '../network/chain-id-to-chain';
+import { PoolController } from '../controllers/pool-controller'; // Add this import statement
 import { sorService } from './sor.service';
 
 describe('sor debugging', () => {
     it('sor v2 arb eth->usdc', async () => {
-        const chain = 'ARBITRUM';
-        const tokenIn = '0x82af49447d8a07e3bd95bd0d56f35241523fbab1';
-        const tokenOut = '0xaf88d065e77c8cc2239327c5edb3a432268e5831';
+        const chain = Chain.FANTOM;
 
+        const chainId = Object.keys(chainIdToChain).find((key) => chainIdToChain[key] === chain) as string;
         initRequestScopedContext();
-        setRequestScopedContextValue('chainId', '42161');
+        setRequestScopedContextValue('chainId', chainId);
         //only do once before starting to debug
         // await poolService.syncAllPoolsFromSubgraph();
         // await poolService.loadOnChainDataForAllPools();
@@ -17,12 +19,42 @@ describe('sor debugging', () => {
 
         const swaps = await sorService.getSorSwapPaths({
             chain,
-            tokenIn,
-            tokenOut,
-            swapType: 'EXACT_OUT',
-            swapAmount: '100',
-            queryBatchSwap: true,
+            tokenIn: '0xf24bcf4d1e507740041c9cfd2dddb29585adce1e', // BAL
+            tokenOut: '0x21be370d5312f44cb42ce377bc9b8a0cef1a4c83', // WETH
+            swapType: 'EXACT_IN',
+            swapAmount: '1',
+            queryBatchSwap: false,
             // useProtocolVersion: 2,
+            // callDataInput: {
+            //     receiver: '0xb5e6b895734409Df411a052195eb4EE7e40d8696',
+            //     sender: '0xb5e6b895734409Df411a052195eb4EE7e40d8696',
+            //     slippagePercentage: '0.1',
+            // },
+        });
+
+        console.log(swaps.returnAmount);
+        expect(parseFloat(swaps.returnAmount)).toBeGreaterThan(0);
+    }, 5000000);
+
+    it('sor v3 sepolia eth->usdc', async () => {
+        const chain = Chain.SEPOLIA;
+
+        const chainId = Object.keys(chainIdToChain).find((key) => chainIdToChain[key] === chain) as string;
+        initRequestScopedContext();
+        setRequestScopedContextValue('chainId', chainId);
+        //only do once before starting to debug
+        await PoolController().addPoolsV3(chainId);
+        await PoolController().syncPoolsV3(chainId);
+        await PoolController().reloadPoolsV3(chain);
+
+        const swaps = await sorService.getSorSwapPaths({
+            chain,
+            tokenIn: '0x8a88124522dbbf1e56352ba3de1d9f78c143751e', // USDC (aave)
+            tokenOut: '0xde46e43f46ff74a23a65ebb0580cbe3dfe684a17', // DAI (aave)
+            swapType: 'EXACT_IN',
+            swapAmount: '0.01',
+            queryBatchSwap: false,
+            useProtocolVersion: 3,
             // callDataInput: {
             //     receiver: '0xb5e6b895734409Df411a052195eb4EE7e40d8696',
             //     sender: '0xb5e6b895734409Df411a052195eb4EE7e40d8696',