+++ modules/sources/transformers/index.ts
@@ -2,3 +2,4 @@ export * from './pool-transformer';
 export * from './pool-tokens-transformer';
 export * from './swap-v3-transformer';
 export * from './tokens-transformer';
+export * from './hook-transformer';