+++ modules/sources/transformers/subgraph-pool-upsert.ts
@@ -1,7 +1,7 @@
 import { Chain } from '@prisma/client';
 import { OnchainPoolData } from '../contracts';
 import { JoinedSubgraphPool } from '../types';
-import { poolTransformer } from './pool-transformer';
+import { poolTransformer, hookTransformer } from './';
 import { formatEther, formatUnits } from 'viem';
 import { poolTokensDynamicDataTransformer, poolTokensTransformer } from './pool-tokens-transformer';
 
@@ -24,6 +24,7 @@ export const subgraphPoolUpsert = (
 
     return {
         pool: poolTransformer(subgraphPoolData, chain),
+        hook: hookTransformer(subgraphPoolData, chain),
         poolDynamicData: {
             id: subgraphPoolData.id,
             poolId: subgraphPoolData.id,
@@ -32,6 +33,8 @@ export const subgraphPoolUpsert = (
             totalSharesNum: Number(formatUnits(onchainPoolData.totalSupply, 18)),
             blockNumber: Number(blockNumber || 0),
             swapFee: formatEther(onchainPoolData.swapFee),
+            aggregateSwapFee: formatEther(onchainPoolData.aggregateSwapFee || 0n),
+            aggregateYieldFee: formatEther(onchainPoolData.aggregateYieldFee || 0n),
             swapEnabled: true,
             holdersCount: Number(subgraphPoolData.holdersCount),
             totalLiquidity: 0,