+++ prisma/migrations/20240808113145_aggregate_fees/migration.sql
@@ -1,0 +1,3 @@
+-- AlterTable
+ALTER TABLE "PrismaPoolDynamicData" ADD COLUMN     "aggregateSwapFee" TEXT NOT NULL DEFAULT '0',
+ADD COLUMN     "aggregateYieldFee" TEXT NOT NULL DEFAULT '0';