+++ test/factories/index.ts
@@ -1,2 +1,6 @@
 export * from './pool.factory';
 export * from './pool_token.factory';
+export * from './prismaPool.factory';
+export * from './prismaPoolDynamicData.factory';
+export * from './prismaToken.factory';
+export * as SOR from './sor';